Add set command
