+++ identity/set.go
@@ -1,4 +1,4 @@
-package cred
+package identity
 
 import (
 	"fmt"
@@ -10,10 +10,12 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// SetCredCmd TODO
-var SetCredCmd = &cobra.Command{
-	Use:   "zit",
-	Short: "git identity manager",
+const dryRunFlag = "dry-run"
+
+// SetCmd is a command that sets git identity based on the configuration file.
+var SetCmd = &cobra.Command{
+	Use:   "set",
+	Short: "Set git identity",
 	Run: func(cmd *cobra.Command, args []string) {
 		ensureGitDir()
 
@@ -53,55 +55,24 @@ defined in the configuration file:
 			cli.PrintlnExit(fmt.Errorf("cannot find a match for host %q", (*repo).Host))
 		}
 
-		cli.PrintlnExit(
-			git.SetConfig("--local", "user.name", cred.name),
-		)
-		cli.PrintlnExit(
-			git.SetConfig("--local", "user.email", cred.email),
-		)
+		dryRun, err := cmd.Flags().GetBool(dryRunFlag)
+		cli.PrintlnExit(err)
+
+		if !dryRun {
+			cli.PrintlnExit(
+				git.SetConfig("--local", "user.name", cred.name),
+			)
+			cli.PrintlnExit(
+				git.SetConfig("--local", "user.email", cred.email),
+			)
+		}
 
 		fmt.Printf("set user: %s <%s>\n", cred.name, cred.email)
 	},
 }
 
-func findBestMatch(conf config.Config, repo git.RepoInfo) (cred *credentials) {
-	if conf.Default != nil {
-		cred = &credentials{
-			conf.Default.Name,
-			conf.Default.Email,
-		}
-	}
-
-	if conf.Overrides != nil {
-		for _, override := range conf.Overrides {
-			if override.Repo != nil {
-				if override.Owner == repo.Owner && *override.Repo == repo.Name {
-					cred = &credentials{
-						override.User.Name,
-						override.User.Email,
-					}
-					break
-				} else {
-					continue
-				}
-			}
-
-			if override.Owner == repo.Owner {
-				cred = &credentials{
-					override.User.Name,
-					override.User.Email,
-				}
-				break
-			}
-		}
-	}
-
-	return
-}
-
-type credentials struct {
-	name  string
-	email string
+func init() {
+	SetCmd.Flags().Bool(dryRunFlag, false, "dry run")
 }
 
 func ensureGitDir() {
