+++ version/version.go
@@ -6,11 +6,13 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// VersionCmd TODO
+const version = "v2.0.0"
+
+// VersionCmd is a command that prints the app version.
 var VersionCmd = &cobra.Command{
 	Use:   "version",
 	Short: "Print version",
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("v1.0.4")
+		fmt.Println(version)
 	},
 }
